+++ extensions/anybox/src/saveCurrentTabWithCollections.tsx
@@ -1,0 +1,5 @@
+import CollectionsForm from "./components/CollectionForm";
+
+export default function SaveCurrentTabWithCollections() {
+  return <CollectionsForm title="Save Current Tab" command="save-current-tab"></CollectionsForm>;
+}