+++ extensions/docsearch/src/vueRouter4.tsx
@@ -1,5 +1,5 @@
 import { SearchDocumentation } from "./components";
 
-export default function Command() {
-  return <SearchDocumentation docsName="Vue Router4" />;
+export default function Command(props: { arguments: { search?: string } }) {
+  return <SearchDocumentation docsName="Vue Router4" quickSearch={props.arguments.search} />;
 }