+++ extensions/docsearch/src/vueuse.tsx
@@ -1,5 +1,5 @@
 import { SearchDocumentation } from "./components";
 
-export default function Command() {
-  return <SearchDocumentation docsName="VueUse" />;
+export default function Command(props: { arguments: { search?: string } }) {
+  return <SearchDocumentation docsName="VueUse" quickSearch={props.arguments.search} />;
 }