+++ extensions/gitfox/src/components/bookmark-list-item.tsx
@@ -1,6 +1,6 @@
-import { Action, ActionPanel, Color, List } from "@raycast/api";
+import { Action, ActionPanel, Color, Icon, List } from "@raycast/api";
 import Bookmark from "../dtos/bookmark-dto";
-import OpenBookMarkAction from "./open-bookmark-action";
+import OpenBookmarkAction from "./open-bookmark-action";
 
 export default function BookmarkListItem(props: { bookmark: Bookmark }) {
   const bookmark: Bookmark = props.bookmark;
@@ -9,25 +9,25 @@ export default function BookmarkListItem(props: { bookmark: Bookmark }) {
     <List.Item
       id={bookmark.RepositoryIdentifier}
       title={bookmark.Name}
-      icon={{ source: "icon-folder.png", tintColor: Color.Blue }}
+      icon={{ source: "repository-30.png", tintColor: Color.Yellow }}
       accessoryTitle={bookmark.getBranch}
-      accessoryIcon={{ source: "icon-branches.png", tintColor: Color.PrimaryText }}
+      accessoryIcon={{ source: "merge-git-30.png", tintColor: Color.SecondaryText }}
       actions={
         <ActionPanel>
           <ActionPanel.Section>
-            <OpenBookMarkAction bookmark={bookmark} />
+            <OpenBookmarkAction bookmark={bookmark} />
             <Action.ShowInFinder path={bookmark.getPath} />
             <Action.OpenWith path={bookmark.getPath} shortcut={{ modifiers: ["cmd", "shift"], key: "return" }} />
             <Action.Open
-              title="Open in Code"
-              icon="icon-vscode.png"
+              title="Open in VSCode"
+              icon="visual-studio-code-30.png"
               target={bookmark.getPath}
               application="Visual Studio Code"
               shortcut={{ modifiers: ["cmd", "shift"], key: "o" }}
             />
             <Action.Open
               title="Open in iTerm"
-              icon="icon-iterm.png"
+              icon={Icon.Terminal}
               target={bookmark.getPath}
               application="iTerm"
               shortcut={{ modifiers: ["cmd", "shift"], key: "t" }}