+++ extensions/google-workspace/src/search-google-drive-files.tsx
@@ -1,16 +1,18 @@
 import { Action, ActionPanel, List, showToast, Toast } from "@raycast/api";
-import { useFetch } from "@raycast/utils";
+import { useFetch, useCachedState } from "@raycast/utils";
 import { useState } from "react";
 
-import { QueryTypes, getFilesURL, File } from "./api/getFiles";
+import { QueryTypes, getFilesURL, File, ScopeTypes } from "./api/getFiles";
 import FileListItem from "./components/FileListItem";
 
 import { withGoogleAuth, getOAuthToken } from "./components/withGoogleAuth";
 
 function SearchGoogleDriveFiles() {
   const [query, setQuery] = useState("");
-  const [queryType, setQueryType] = useState<QueryTypes>(QueryTypes.fileName);
-  const { data, isLoading, mutate } = useFetch<{ files: File[] }>(getFilesURL(queryType, query), {
+  const [queryType, setQueryType] = useCachedState<QueryTypes>("query type", QueryTypes.fileName);
+  const [scopeType, setScopeType] = useCachedState<ScopeTypes>("scope type", ScopeTypes.allDrives);
+
+  const { data, isLoading, mutate } = useFetch<{ files: File[] }>(getFilesURL(queryType, scopeType, query), {
     headers: {
       "Content-Type": "application/json",
       Authorization: `Bearer ${getOAuthToken()}`,
@@ -26,9 +28,25 @@ function SearchGoogleDriveFiles() {
       isLoading={isLoading}
       searchBarPlaceholder="Search files"
       searchBarAccessory={
-        <List.Dropdown tooltip="Search mode" storeValue onChange={setQueryType as (value: string) => void}>
-          <List.Dropdown.Item title="By file name" value={QueryTypes.fileName} />
-          <List.Dropdown.Item title="In full text" value={QueryTypes.fullText} />
+        <List.Dropdown
+          tooltip="Search mode"
+          value={`${queryType}-${scopeType}`}
+          onChange={(value) => {
+            const [queryType, scopeType] = value.split("-");
+            setQueryType(queryType as QueryTypes);
+            setScopeType(scopeType as ScopeTypes);
+          }}
+        >
+          <List.Dropdown.Item title="By file name in My Drive" value={`${QueryTypes.fileName}-${ScopeTypes.user}`} />
+          <List.Dropdown.Item
+            title="By file name in All Drives"
+            value={`${QueryTypes.fileName}-${ScopeTypes.allDrives}`}
+          />
+          <List.Dropdown.Item title="In full text in My Drive" value={`${QueryTypes.fullText}-${ScopeTypes.user}`} />
+          <List.Dropdown.Item
+            title="In full text in All Drives"
+            value={`${QueryTypes.fullText}-${ScopeTypes.allDrives}`}
+          />
         </List.Dropdown>
       }
       onSearchTextChange={setQuery}