+++ extensions/itranslate/CHANGELOG.md
@@ -1,5 +1,20 @@
 # iTranslate Changelog
 
+## [Add New Translation Service Providers] - {PR MERGE DATE}
+
+- Add Microsoft Azure option
+
+## [Translation History & TTS] - 2022-09-02
+
+- Add new command `Translation History`. Support record translation history
+- Support TTS by OS command `say`
+
+## [Add New Translation Service Providers & New icon] - 2022-08-26
+
+- Add Google Could Translation option
+- Add Aliyun option
+- New icon
+
 ## [Optimize Experience] - 2022-08-18
 
 - Replace List.throttle with setTimeout to optimize the user input experience